+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Hard fail if any config option requires `ffmpeg` but `ffmpeg` is not installed
+
 ### Deprecated
 
 ### Removed